requirements.txt
aws-cdk-lib==2.133.0
constructs>=10.0.0,<11.0.0
h5py
typing-extensions
wheel
aws-cdk.aws_kinesisanalytics_flink_alpha